[Refactor] Service는 service로 옮기기
